convert execbarval to double

+++ src/exec.h
@@ -42,7 +42,7 @@ void print_execp(struct text_object *, char *, int);
 void print_execi(struct text_object *, char *, int);
 void print_execpi(struct text_object *, char *, int);
 void print_texeci(struct text_object *, char *, int);
-uint8_t execbarval(struct text_object *);
+double execbarval(struct text_object *);
 uint8_t execi_barval(struct text_object *);
 void free_exec(struct text_object *);
 void free_execi(struct text_object *);
